Applicationクラス関係のコミット。同時にその他いろいろやりすぎてる

+++ Hidder/Models/Application.cs
@@ -1,34 +1,43 @@
-﻿using Hidder.Common;
+﻿using Hidder.Models;
 using System;
 using System.Diagnostics;
 using System.IO;
-using System.Runtime.InteropServices;
 using System.Windows.Forms;
+using System.Xml.Serialization;
 
 namespace Hidder.Models
 {
-    public class ApplicationProcess : PropertyChangeNotification
+    /// <summary>
+    /// 実行アプリケーションクラス。
+    /// </summary>
+    [Serializable]
+    [XmlRoot("Application")]
+    public class Application : PropertyChangeNotification
     {
         private System.Diagnostics.Process _process;
 
         /// <summary>
         /// ID (データグリッドと整合性とるためのIDでプロセスIDじゃない)
+        /// Hack:xmlからリストアする場合に外部からIDを指定するためにPublicにする。
         /// </summary>
-        public int Id { get; private set; }
+        [XmlIgnore]
+        public int Id { get; set; }
 
         #region"プロセスに関する情報"
 
         private bool _isRunning = false;
         /// <summary>
         /// プロセス稼動有無
+        /// Hack:Dispose済みかどうか確認するためにも使用する。_processの生成/破棄と同期を取る。
         /// </summary>
+        [XmlIgnore]
         public bool IsRunning
         {
             get { return this._isRunning; }
             private set
             {
                 this._isRunning = value;
-                if(value)
+                if (value)
                 {
                     this.ApplicationStatus = ApplicationStatusEnum.Running;
                 }
@@ -40,6 +49,7 @@ namespace Hidder.Models
         /// <summary>
         /// プロセス稼動詳細ステータス
         /// </summary>
+        [XmlIgnore]
         public ApplicationStatusEnum ApplicationStatus
         {
             get { return this._applicationStatus; }
@@ -50,11 +60,12 @@ namespace Hidder.Models
             }
         }
 
-        private ProcessWindowStyle _currentWindowStyle;
+        private WindowStyleEnum _currentWindowStyle;
         /// <summary>
         /// 現在のウインドウサイズ
         /// </summary>
-        public ProcessWindowStyle CurrentWindowStyle
+        [XmlIgnore]
+        public WindowStyleEnum CurrentWindowStyle
         {
             get { return this._currentWindowStyle; }
             private set
@@ -68,6 +79,7 @@ namespace Hidder.Models
         /// <summary>
         /// プロセスID
         /// </summary>
+        [XmlIgnore]
         public string ProcessId
         {
             get { return this._processId; }
@@ -78,37 +90,74 @@ namespace Hidder.Models
             }
         }
 
+        private string _title = "";
+        /// <summary>
+        /// タイトル
+        /// </summary>
+        [XmlElement]
+        public string Title
+        {
+            get { return this._title; }
+            set
+            {
+                this._title = value;
+                this.OnPropertyChanged();
+            }
+        }
+
         /// <summary>
         /// パス
         /// </summary>
-        public string FullPath { get; private set; }
+        [XmlElement]
+        public string FullPath { get; set; }
 
         /// <summary>
         /// ファイル名
         /// </summary>
+        [XmlIgnore]
         public string FileName { get { return Path.GetFileName(this.FullPath); } }
 
+        private string _argument = "";
         /// <summary>
         /// 引数
         /// </summary>
-        public string Argument { get; private set; }
+        [XmlElement]
+        public string Argument
+        {
+            get { return this._argument; }
+            set
+            {
+                this._argument = value;
+                this.OnPropertyChanged();
+            }
+        }
 
         /// <summary>
-        /// GUIかCUIか
-        /// Fixme:Enumにする。
+        /// アプリケーションのUI種別
         /// </summary>
-        private string UIType { get; set; }
+        [XmlIgnore]
+        private ApplicationUiType UiType { get; set; }
 
         #endregion
 
         /// <summary>
         /// コンストラクタ
         /// </summary>
-        public ApplicationProcess(int id, string path, string argument)
+        public Application()
         {
+            //シリアライズのために引数の存在しないコンストラクタが必要
+        }
+
+        /// <summary>
+        /// コンストラクタ
+        /// </summary>
+        public Application(int id, string title, string path, string argument)
+        {
+            this.Title = title;
             this.Id = id;
             this.FullPath = path;
             this.Argument = argument;
+            this.CurrentWindowStyle = WindowStyleEnum.None;
             if (!System.IO.File.Exists(path))
             {
                 this.ApplicationStatus = ApplicationStatusEnum.DoesntExist;
@@ -125,6 +174,7 @@ namespace Hidder.Models
             //Hack:ここでTry~Catchしたくないんだけどな～
             try
             {
+                this.IsRunning = true;
                 this.ApplicationStatus = ApplicationStatusEnum.Starting;
                 _process = new System.Diagnostics.Process();
                 _process.StartInfo.FileName = path;
@@ -137,21 +187,22 @@ namespace Hidder.Models
                 System.Threading.Thread.Sleep(500);
 
                 //Hack:InternalNameがcmdかどうかでCUIかGUIを判断する。
-                if (_process.MainModule.FileVersionInfo.InternalName != "cmd")
+                if (_process.MainModule.FileVersionInfo.InternalName == "cmd")
                 {
-                    this.UIType = "GUI";
-                    _process.WaitForInputIdle(10000);
+                    this.UiType = ApplicationUiType.CUI;
                 }
                 else
                 {
-                    this.UIType = "CUI";
+                    this.UiType = ApplicationUiType.GUI;
+                    //アプリケーションが入力可能になるまで最大10秒待機する
+                    _process.WaitForInputIdle(10000);
                 }
 
-                this.CurrentWindowStyle = _process.StartInfo.WindowStyle;
+                this.CurrentWindowStyle = WindowStyleEnum.Normal;
                 this.FullPath = path;
                 this.Argument = argument;
                 this.ProcessId = _process.Id.ToString();
-                this.IsRunning = true;
+                this.ApplicationStatus = ApplicationStatusEnum.Running;
             }
             catch (Exception ex)
             {
@@ -166,7 +217,7 @@ namespace Hidder.Models
         /// </summary>
         public void Show()
         {
-            if (this.CurrentWindowStyle == ProcessWindowStyle.Hidden)
+            if (this.CurrentWindowStyle == WindowStyleEnum.Hidden)
             {
                 this.CurrentWindowStyle = WindowStyleChange.Normal(this._process);
             }
@@ -177,7 +228,7 @@ namespace Hidder.Models
         /// </summary>
         public void Hide()
         {
-            if (this.CurrentWindowStyle != ProcessWindowStyle.Hidden)
+            if (this.CurrentWindowStyle != WindowStyleEnum.Hidden)
             {
                 this.CurrentWindowStyle = WindowStyleChange.Hide(this._process);
             }
@@ -185,45 +236,59 @@ namespace Hidder.Models
 
         /// <summary>
         /// プロセスを終了します。
-        /// Fixme:ここの例外処理考える。
         /// </summary>
         public void Kill()
         {
             this.ApplicationStatus = ApplicationStatusEnum.Stopping;
+
+            if (MessageBox.Show("Do you really want to kill it process force ?", "Warning"
+                , MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation, MessageBoxDefaultButton.Button2) == DialogResult.No)
+            {
+                this.ApplicationStatus = ApplicationStatusEnum.Running;
+                return;
+            }
+
             try
             {
-                if (MessageBox.Show("Do you really want to kill it process force ?", "Warning"
-                   , MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation, MessageBoxDefaultButton.Button2) == DialogResult.No)
+                //メッセージボックス確認中にプロセスが終了している可能性があるので、確認後にKillする。
+                if (this.IsRunning == true) _process.CloseMainWindow();
+
+                if (UiType == ApplicationUiType.GUI)
                 {
-                    this.ApplicationStatus = ApplicationStatusEnum.Running;
-                    return;
+                    KillGui();
                 }
-
-                _process.CloseMainWindow();
-
-                if (UIType == "CUI")
-                {
-                    KillProcessFinalize();
-                    return;
-                }
-
-                if (!_process.HasExited)
-                {
-                    //GUIアプリケーションが終了しなかった場合、強制終了させるか再度確認する。
-                    if (MessageBox.Show("Do you really want to kill GUI application force ?", "Warning"
-                        , MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation, MessageBoxDefaultButton.Button2) == DialogResult.No)
-                    {
-                        this.ApplicationStatus = ApplicationStatusEnum.Running;
-                        return;
-                    }
-                    _process.Kill();
-                };
-                KillProcessFinalize();
             }
             catch (Exception ex)
             {
                 System.Windows.MessageBox.Show((string)ex.Message + Environment.NewLine + ex.StackTrace);
                 this.ApplicationStatus = ApplicationStatusEnum.Faild;
+            }
+            finally
+            {
+                //IsRunningでプロセスが終了しているかどうか判断する。
+                if (this.IsRunning == true) KillProcessFinalize();
+            }
+        }
+
+        /// <summary>
+        /// GUIアプリケーションのプロセスを終了します。
+        /// </summary>
+        private void KillGui()
+        {
+            //アプリケーションに終了メッセージ送信後に最大10秒待機する
+            System.Threading.Thread.Sleep(10000);
+
+            if (!_process.HasExited)
+            {
+                //GUIアプリケーションが終了しなかった場合、強制終了させるか再度確認する。
+                if (MessageBox.Show("Do you really want to kill GUI application force ?", "Warning"
+                    , MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation, MessageBoxDefaultButton.Button2) == DialogResult.No)
+                {
+                    this.ApplicationStatus = ApplicationStatusEnum.Running;
+                    return;
+                }
+                //メッセージボックス確認中にプロセスが終了している可能性があるので、確認後にKillする。
+                if (this.IsRunning == true) _process.Kill();
             }
         }
 
@@ -245,9 +310,9 @@ namespace Hidder.Models
             this.IsRunning = false;
             this.ApplicationStatus = ApplicationStatusEnum.Ready;
             this.ProcessId = "-";
+            this.CurrentWindowStyle = WindowStyleEnum.None;
             _process.Close();
             _process.Dispose();
         }
-
     }
 }
